requirements.txt
discord.py==2.4.0
google-genai==0.6.0
python-dotenv==1.0.1
asyncpg==0.30.0
openai==1.55.3
aioconsole==0.8.1
